# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WChat 即时通讯系统后端 API 文档",
        "title": "WChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户收到的所有好友/群聊申请，包含申请人昵称、头像等信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "申请"
                ],
                "summary": "获取收到的申请列表",
                "responses": {
                    "200": {
                        "description": "Token 无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向目标用户发送好友申请(contact_type=0)或向群聊发送加群申请(contact_type=1)，可附加留言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "申请"
                ],
                "summary": "发起好友/群聊申请",
                "parameters": [
                    {
                        "description": "申请参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SubmitApplicationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "不能添加自己 / 已经是好友 / 申请处理中",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/applications/{uuid}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对收到的申请进行处理：1=同意（自动建立双向好友关系）、2=拒绝、3=拉黑",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "申请"
                ],
                "summary": "处理申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "申请记录 UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "处理操作: 1=同意, 2=拒绝, 3=拉黑",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.HandleApplicationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "申请不存在 / 权限不足 / 已处理过",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "通过手机号和密码登录，验证成功后下发 JWT Token 及用户基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "账号或密码错误 / 账号被禁用",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登出当前账号，更新最后离线时间（后续可扩展 Token 黑名单机制）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "Token 无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "通过手机号、密码和昵称注册新账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "鉴权"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 手机号已注册",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.HandleApplicationReq": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "description": "1:同意 2:拒绝 3:拉黑",
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                }
            }
        },
...
        "types.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "telephone"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "types.RegisterReq": {
            "type": "object",
            "required": [
                "confirm_password",
                "nickname",
                "password",
                "telephone"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 20
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "telephone": {
                    "description": "根据你的表结构，使用 Telephone 作为主账号",
                    "type": "string"
                }
            }
        },
...
        "types.SubmitApplicationReq": {
            "type": "object",
            "required": [
                "contact_id"
            ],
            "properties": {
                "contact_id": {
                    "type": "string"
                },
                "contact_type": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "message": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}